monitor/requirements.txt
altgraph==0.16.1
configparser==3.7.4
future==0.17.1
pefile==2019.4.18
psutil==5.6.3
pywin32==224
pywin32-ctypes==0.2.0
